Fixed bugs in SHVC playback support
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@5177 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/media_tools/html5_mse.c
@@ -641,7 +641,7 @@ static GF_Err gf_mse_process_coded_frame(GF_HTML_SourceBuffer    *sb,
     }
 
     if (track->last_dts_set) {
-        if (DTS_with_offset < track->last_dts ||
+        if (DTS_with_offset < (s64) track->last_dts ||
 			DTS_with_offset - track->last_dts > 2*track->last_dur) {
 			/* A discontinuity in the timestamps is detected, this triggers the start of a new coded frame group */
 			if (sb->append_mode == MEDIA_SOURCE_APPEND_MODE_SEGMENTS) {
